requirements.txt
beautifulsoup4==4.12.3
python-telegram-bot==21.1.1
